Split generic annotation diagnostics
